+++ src/plone/api/tests/test_content.py
@@ -5,21 +5,27 @@ from Acquisition import aq_base
 from OFS.CopySupport import CopyError
 from OFS.event import ObjectWillBeMovedEvent
 from OFS.interfaces import IObjectWillBeMovedEvent
-from Products.CMFCore.interfaces import IContentish
 from Products.CMFCore.WorkflowCore import WorkflowException
+from Products.CMFCore.interfaces import IContentish
 from Products.ZCatalog.interfaces import IZCatalog
 from plone import api
+from plone.api.content import NEW_LINKINTEGRITY
 from plone.api.tests.base import INTEGRATION_TESTING
+from plone.app.linkintegrity.exceptions import \
+    LinkIntegrityNotificationException
+from plone.app.textfield import RichTextValue
+from plone.dexterity.interfaces import IDexterityContent
 from plone.indexer import indexer
 from plone.uuid.interfaces import IMutableUUID
 from plone.uuid.interfaces import IUUIDGenerator
 from zExceptions import BadRequest
+from zope.component import getGlobalSiteManager
+from zope.component import getUtility
+from zope.container.contained import ContainerModifiedEvent
 from zope.lifecycleevent import IObjectModifiedEvent
 from zope.lifecycleevent import IObjectMovedEvent
 from zope.lifecycleevent import ObjectMovedEvent
-from zope.component import getUtility
-from zope.component import getGlobalSiteManager
-from zope.container.contained import ContainerModifiedEvent
+from zope.lifecycleevent import modified
 
 import mock
 import pkg_resources
@@ -596,7 +602,7 @@ class TestPloneApiContent(unittest.TestCase):
 
         # Delete the contact page
         api.content.delete(self.contact)
-        assert 'contact' not in container['about'].keys()
+        self.assertNotIn('contact', container['about'].keys())
 
     def test_delete_multiple(self):
         """Test deleting multiple content items."""
@@ -607,8 +613,97 @@ class TestPloneApiContent(unittest.TestCase):
 
         api.content.delete(objects=[container['copy_of_about'],
                                     container['events']['about']])
-        assert 'copy_of_about' not in container
-        assert 'about' not in container['events']
+        self.assertNotIn('copy_of_about', container)
+        self.assertNotIn('about', container['events'])
+
+    def test_delete_ignore_linkintegrity(self):
+        """Test deleting a content item with a link pointed at it."""
+        self._set_text(self.team, '<a href="contact">contact</a>')
+        # Delete the contact page
+        api.content.delete(self.contact, check_linkintegrity=False)
+        self.assertNotIn('contact', self.portal['about'].keys())
+
+    @unittest.skipIf(
+        HAS_PACONTENTYPES and not NEW_LINKINTEGRITY,
+        'This test only makes sense with Archetypes or new Linkintegrity.')
+    def test_delete_check_linkintegrity(self):
+        """Test deleting a content item with a link pointed at it."""
+        self._set_text(self.team, '<a href="contact">contact</a>')
+        # Delete the contact page
+        with self.assertRaises(LinkIntegrityNotificationException):
+            api.content.delete(self.contact)
+        if NEW_LINKINTEGRITY:
+            # In the old implementation of linkintegrity the items are
+            # still gone during this request.
+            self.assertIn('contact', self.portal['about'].keys())
+
+    @unittest.skipIf(
+        HAS_PACONTENTYPES and not NEW_LINKINTEGRITY,
+        'This test only makes sense with Archetypes or new Linkintegrity.')
+    def test_delete_multiple_check_linkintegrity(self):
+        """Test deleting multiple item with linkintegrity-breaches."""
+        self._set_text(self.team, '<a href="../about/contact">contact</a>')
+        self._set_text(self.training, '<a href="../blog">contact</a>')
+        # Delete the contact page
+        with self.assertRaises(LinkIntegrityNotificationException):
+            api.content.delete(objects=[self.blog, self.contact])
+        if NEW_LINKINTEGRITY:
+            # In the old implementation of linkintegrity the items are
+            # still gone during this request.
+            self.assertIn('contact', self.portal['about'].keys())
+            self.assertIn('blog', self.portal.keys())
+
+    @unittest.skipIf(
+        HAS_PACONTENTYPES and not NEW_LINKINTEGRITY,
+        'This test only makes sense with Archetypes or new Linkintegrity.')
+    def test_delete_multiple_ignore_linkintegrity(self):
+        """Test deleting multiple items ignoring linkintegrity-breaches."""
+        self._set_text(self.team, '<a href="../about/contact">contact</a>')
+        self._set_text(self.training, '<a href="../blog">contact</a>')
+        # Delete linked pages
+        api.content.delete(
+            objects=[self.blog, self.contact],
+            check_linkintegrity=False)
+        self.assertNotIn('contact', self.portal['about'].keys())
+        self.assertNotIn('blog', self.portal.keys())
+
+    @unittest.skipIf(
+        HAS_PACONTENTYPES and not NEW_LINKINTEGRITY,
+        'This test only makes sense with Archetypes or new Linkintegrity.')
+    def test_delete_with_internal_breaches(self):
+        """Test deleting multiple with internal linkintegrity breaches."""
+        self._set_text(self.team, '<a href="../about/contact">contact</a>')
+        self._set_text(self.training, '<a href="../blog">contact</a>')
+        # Deleting pages with unresolved breaches throws an exception
+        with self.assertRaises(LinkIntegrityNotificationException):
+            api.content.delete(objects=[self.blog, self.about])
+        if NEW_LINKINTEGRITY:
+            # In the old implementation of linkintegrity the items are
+            # still gone during this request.
+            self.assertIn('about', self.portal.keys())
+            self.assertIn('blog', self.portal.keys())
+            self.assertIn('training', self.portal['events'].keys())
+
+    @unittest.skipUnless(
+        NEW_LINKINTEGRITY, 'Only new Linkintegrity resolves internal breaches')
+    def test_delete_with_resolved_internal_breaches(self):
+        """Test deleting multiple with internal linkintegrity breaches."""
+        self._set_text(self.team, '<a href="../about/contact">contact</a>')
+        self._set_text(self.training, '<a href="../blog">contact</a>')
+        # Deleting pages with resolved breaches throws no exception
+        api.content.delete(objects=[self.blog, self.training, self.about])
+        self.assertNotIn('about', self.portal.keys())
+        self.assertNotIn('blog', self.portal.keys())
+        self.assertNotIn('training', self.portal['events'].keys())
+
+    def _set_text(self, obj, text):
+        if IDexterityContent.providedBy(obj):
+            # Dexterity
+            obj.text = RichTextValue(text)
+        else:
+            # Archetypes
+            obj.setText(text, mimetype='text/html')
+        modified(obj)
 
     def test_find(self):
         """Test the finding of content in various ways."""